feat: add color format selection for `ColorInput`
